remove rechart
