+++ package.json
@@ -14,7 +14,6 @@
     "dependencies": {
         "nodemon": "^2.0.2",
         "split2": "3.1.1",
-        "recharts": "1.8.5",
         "@material-ui/core": "^4.10.0",
         "@material-ui/styles": "^4.10.0"
     },
